+++ packaging/nix/flake.nix
@@ -13,7 +13,7 @@
       in {
         packages.default = pkgs.rustPlatform.buildRustPackage {
           pname = "fono";
-          version = "0.1.0";
+          version = "0.2.0";
           src = ./.;
 
           cargoLock = {